updated float instructions to use single precision. added rv64_f and rv64_d, yet to implement instructions

+++ sim/src/insn_impl/fcvt_s_w.rs
@@ -5,7 +5,7 @@ pub fn fcvt_s_w(insn: Insn, cpu: &mut Cpu, bus: &mut Bus) -> cpu::Result<u64> {
     let rs1 = insn.rs1();
     let rm = insn.rm();
 
-    let value = cpu.fload(rs1) as f32 as f64;
+    let value = cpu.fload(rs1) as i32 as f32 as f64;
     cpu.fstore(rd, value);
     Ok(cpu.pc + 4)
 }
