+++ sim/src/insn_impl/fcvt_wu_s.rs
@@ -5,7 +5,7 @@ pub fn fcvt_wu_s(insn: Insn, cpu: &mut Cpu, bus: &mut Bus) -> cpu::Result<u64> {
     let rs1 = insn.rs1();
     let rm = insn.rm();
 
-    let value = cpu.fload(rs1) as u32 as f64;
+    let value = cpu.fload(rs1) as f32 as u32 as f64;
     cpu.fstore(rd, value);
     Ok(cpu.pc + 4)
 }
