+++ sim/src/insn_impl/fle_s.rs
@@ -5,7 +5,7 @@ pub fn fle_s(insn: Insn, cpu: &mut Cpu, bus: &mut Bus) -> cpu::Result<u64> {
     let rs1 = insn.rs1();
     let rs2 = insn.rs2();
 
-    let value = if cpu.fload(rs1) <= cpu.fload(rs2) {
+    let value = if cpu.fload(rs1) as f32 <= cpu.fload(rs2) as f32 {
         1.0
     } else {
         0.0
