+++ sim/src/insn_impl/fsqrt_s.rs
@@ -5,7 +5,7 @@ pub fn fsqrt_s(insn: Insn, cpu: &mut Cpu, bus: &mut Bus) -> cpu::Result<u64> {
     let rs1 = insn.rs1();
     let rm = insn.rm();
 
-    let value = cpu.fload(rs1).sqrt();
+    let value = (cpu.fload(rs1) as f32).sqrt() as f64;
     cpu.fstore(rd, value);
     Ok(cpu.pc + 4)
 }
